+++ src/frontend/src/components/Carousels/CarouselLargePartnerCard/index.jsx
