+++ src/frontend/src/components/Carousels/CarouselLargePartnerCard/styles.module.scss
